requirements.txt
dataclasses==0.6
dataclasses-json==0.5.6
pyserial==3.5
